+++ CHANGELOG.rst
@@ -2,6 +2,12 @@
 Change Log
 ==========
 
+``master`` (unreleased)
+
+* Allow to start Black Mamba even in untested version of Pythonista, just
+  warn the user
+ 
+
 ``0.0.15`` (2017-08-28)
 -----------------------